Add support for 32-bit pointers under the special 64-32 interop environment.
With this, it's now possible to declare 32-bit pointers using the
`__ptr32` syntax under the 64-32 interop (i.e. `wine32`) environment.
These pointers can be dereferenced and used like any other. Function
pointers can even be called. A 32-bit pointer can additionally be
expanded to a 64-bit one, but not the reverse.

This also puts all stack variables on this target in that address space.
That means that, by default, taking the address of a stack variable
yields a `__ptr32` pointer. This is necessary to be able to pass these
addresses to functions expecting a 32-bit pointer.