Add a pragma to set the default address space.
Currently, this pragma can be used to set the default address space for
pointers, references, and any global data to any target address space.
Additonally, in 64-32 interop mode, it can be used to enable 32-bit
pointers by default.  Unfortunately, OpenCL and CUDA address spaces
aren't supported yet--I don't understand how they work. This may require
additional work to implement this correctly for those two environments.

This pragma supports MSVC-style stack operations. Push, pop, and reset
are all supported. This means you can do:

```c++
  #pragma clang default_addr_space(push, ptr32)
  #include <windows.h>
  #pragma clang default_addr_space(pop)
```

and all the declarations in `<windows.h>` will automatically be in the
32-bit address space, while those that follow won't be.

One side effect of this change is that string literals are now emitted
in the address space specified by this pragma--unless the target
specifies that they belong in a particular AS. This is necessary to
avoid a crash later on in LLVM when passing a string literal to a
`__ptr32` parameter.