Add support for variadic 32-bit ABI functions in 64-bit mode.
The new type `__builtin_va_list32` provides access to varargs from
variadic 32-bit ABI functions (i.e. those marked `cdecl32` with a `...`
in their argument lists). It's used in the same way as a normal
`va_list`, but it obeys the semantics and rules of 32-bit argument
passing. The corresponding `va_start()`, `va_end()`, and `va_copy()`
routines are suffixed with a `32`.