Add an option to set the default address space.
This can be used to set the default address space from the command line.
You can override it using the `clang default_addr_space()` pragma.
Saying:

```c++
  #pragma clang default_addr_space()
```

will cause the address space to revert to this setting.

If `-mwine32` is passed, `-mdefault-address-space=ptr32` is now implied.
To override this, pass an explicit `-mdefault-address-space` option.