[CodeGen] Emit references to 32-bit-callable functions' thunks.
Do this for any reference to a 32-bit-callable function other than a
direct call. This way, code can store to a 32-bit function pointer a
32-bit-calling-convention function, and the pointer will magically
have the correct value.

Unfortunately, it does this even for references that would result in a
64-bit pointer. I had hoped to be able to make a 64-bit reference
automatically refer to the real function instead of the thunk, but that
may need to wait until I can think up a better solution for this.