[AST] Put pointers to system types in the system AS.
"System types" here means those types defined in a system header.
Currently, this is limited to `typedef`s, structures, unions, classes,
and enumerations. The assumption is that pointers to such types,
particularly those consumed or returned from system functions, are
likely to be in the system AS. A good example is the `FILE` type from
the `stdio` package. Pointers to it are returned by functions such as
`fopen(3)` in the system C library. If `fopen(3)` is declared in a
system header, it will return a pointer in the system address space. But
a naive application is likely to declare a variable to hold the file
pointer like this:

```c
FILE *f = fopen("foo.txt", "w+");
```

If we used the default AS, it is not the same as the system AS, and it
is not a superset of the system AS, this would fail before. Now it
won't.