[Sema] Add a kludge to make va_list parameters work with va_arg(3).
In 64-32 interop mode, `va_arg(3)` for System V wants a
`struct __va_list_tag* __ptr32`, which a `va_list` declared on the stack
will normally satisfy. But a `va_list` passed as a parameter won't
produce that; it produces a `__ptr64` pointer instead. This is because
the `va_list` type is really an array, which decays to a pointer in the
parameter list. But by the time this happens, information on the address
space has been lost. So, we need to accept both `__ptr32` and `__ptr64`
`va_list`s for `va_arg(3)` to work.