[CodeGen] Cast array-to-pointer-decay expressions to the correct AS.
If a GNU compound array literal is passed for a parameter pointer to the
global AS, but the temporary used to hold it is allocated on the stack,
and the stack AS differs from the global AS, then the address spaces of
the parameter and the decayed pointer won't match. Make them match up
when we emit the pointer expression.